+++ NEWS.md
@@ -1,6 +1,13 @@
 
 # renv 0.18.0  (UNRELEASED)
 
+* `dependencies()` now marks `Suggested` packages listed in `DESCRIPTION` files
+  as development dependencies regardless of whether or not they're a "package" 
+  project.
+
+* `settings$package.dependency.fields()` now only affects packages installed
+  directly by the user, not downstream dependencies of those packages.
+
 * Fixed an issue where `renv::snapshot(exclude = <...>)` could warn when
   attempting to exclude a package which was not already installed. (#1396)